llama_adapter_v2_multimodal7b/requirements.txt
aioboto3==14.1.0
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.37.1
botocore==1.37.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clip==1.0
cmake==4.0.0
deeplake==4.1.19
dill==0.4.0
exceptiongroup==1.2.2
fairscale==0.4.13
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.3.1
gradio==5.25.1
gradio_client==1.8.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humbug==0.3.2
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
libdeeplake==0.0.162
lit==18.1.8
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathos==0.3.4
pillow==11.2.1
pox==0.3.6
ppft==1.7.7
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.5
s3transfer==0.11.3
safehttpx==0.1.6
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
threadpoolctl==3.6.0
timm==0.3.2
tomlkit==0.13.2
torch==2.0.0+cu117
torchvision==0.15.1+cu117
tqdm==4.67.1
triton==2.0.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
